no message
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@441 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ modules/dx_hw/dx_window.c
@@ -332,7 +332,7 @@ static void SetWindowless(GF_VideoOutput *vout, HWND hWnd)
 	sscanf(opt, "%02X%02X%02X%02X", &a, &r, &g, &b);
 	ckey = RGB(r, g, b);
 	if (a<255)
-		_SetLayeredWindowAttributes(hWnd, ckey, a, LWA_COLORKEY|LWA_ALPHA);
+		_SetLayeredWindowAttributes(hWnd, ckey, (u8) a, LWA_COLORKEY|LWA_ALPHA);
 	else
 		_SetLayeredWindowAttributes(hWnd, ckey, 0, LWA_COLORKEY);
